services/student/requirements.txt
alembic==1.0.7
amqp==2.4.1
astroid==2.1.0
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
dnspython==1.16.0
eventlet==0.24.1
greenlet==0.4.15
idna==2.8
importlib-metadata==0.8
isort==4.3.4
kombu==4.3.0
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
mock==2.0.0
monotonic==1.5
more-itertools==6.0.0
nameko==2.11.0
nameko-sqlalchemy==1.4.0
passlib==1.7.1
path.py==11.5.0
pbr==5.1.2
pg8000==1.13.1
pluggy==0.8.1
psycopg2==2.7.7
py==1.8.0
pylint==2.2.2
pytest==4.3.0
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==3.13
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.18
testing.common.database==2.0.3
testing.postgresql==1.3.0
urllib3==1.24.1
vine==1.2.0
Werkzeug==0.14.1
wrapt==1.11.1
zipp==0.3.3
